adjust the parameters of the flippant.flip function for better control over the back element.
Changed the parameter `content` to the HTML element--that is the back of the card/modal--with the name `back`.

The older way was flawed in the sense that, when the `back` element was created, it was being appended to the `body` hence reducing the control over the flippand.

To better understand this, imagine a scenario where the developer wants to destroy the flipped card (i.e. the `back` element) when the user requests a new view (the Backbone-way). With the older way, it was incredibly hard to destroy the back element.

If the developer specifies the `back` element by himself (as this commit allows), s/he can destroy the container of the card/modals and will simply get rid of the overflowing back element whenever a new view is created.

Just a suggestion ;)